pyproject.toml
[project]
name = "emostack"
version = "0.1.0"
description = ""
authors = [
    {name = "Eliza Marszalik"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "gradio (>=5.34.1,<6.0.0)",
    "openai (>=1.88.0,<2.0.0)",
    "sqlite-web (>=0.6.4,<0.7.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
